### Description This pull request adds authentication and authorization treatments for routes
### Changes included
- category-routes.tsx

+++ src/app/logout/page.tsx
@@ -1,19 +1,35 @@
 'use client';
-import { tokenService } from '@/services/Auth/tokenService';
+
 import { useRouter } from 'next/navigation';
-import React from 'react';
+import { useEffect, useRef } from 'react';
+import { toast } from 'react-toastify';
+import { tokenService } from '@/services/Auth/tokenService';
 
 export default function LogoutPage() {
   const router = useRouter();
+  const hasLoggedOut = useRef(false);
 
-  React.useEffect(() => {
-    try {
-      tokenService.delete();
-      router.push('/');
-    } catch (error) {
-      alert(error);
-    }
+  useEffect(() => {
+    const logout = async () => {
+      if (hasLoggedOut.current) return;
+
+      hasLoggedOut.current = true;
+      try {
+        await tokenService.delete();
+        sessionStorage.removeItem('loginStatus');
+        toast.info('You are being logged out...');
+        const timer = setTimeout(() => {
+          router.push('/');
+        }, 2000);
+
+        return () => clearTimeout(timer);
+      } catch (error) {
+        toast.error(`Logout failed: ${(error as Error).message}`);
+      }
+    };
+
+    logout();
   }, [router]);
 
-  return <div>Logging out...</div>;
+  return null;
 }
